p4 version of shell
Tested with wc.  The program was found but execve overrides the entire program with no output. I'm wondering if the redirect structure of the original p4 has something to do with it, meaning that the output is being put into the wrong fd or that the output is not even turned on for the original p4 program.