Finally got execve call to work!
This version, as the filename indicates, is based off of the p3-exec.py file in the original os-demo repo.  I got the execve() call to execute the program passed in with its arguments.  For some reason, when calling wc <file> I have to hit ctrl-d to get a print out of counts.  I also tried the execve() call with grep.  Grep runs and then ends, control goes back to parent, which ends shortly after.